LPS-53856 elasticsearch-rename rename
