+++ modules/portal/portal-search-elasticsearch/src/com/liferay/portal/search/elasticsearch/ElasticsearchSpellCheckIndexWriter.java
@@ -24,7 +24,6 @@ import com.liferay.portal.kernel.search.SearchContext;
 import com.liferay.portal.kernel.search.SearchException;
 import com.liferay.portal.kernel.search.SpellCheckIndexWriter;
 import com.liferay.portal.search.elasticsearch.connection.ElasticsearchConnectionManager;
-import com.liferay.portal.search.elasticsearch.spi.ElasticsearchUpdateDocumentCommand;
 import com.liferay.portal.search.elasticsearch.util.DocumentTypes;
 import com.liferay.portal.search.elasticsearch.util.LogUtil;
