+++ modules/portal/portal-search-elasticsearch/src/com/liferay/portal/search/elasticsearch/facet/CompositeFacetProcessor.java
@@ -16,7 +16,6 @@ package com.liferay.portal.search.elasticsearch.facet;
 
 import com.liferay.portal.kernel.search.facet.Facet;
 import com.liferay.portal.kernel.util.MapUtil;
-import com.liferay.portal.search.elasticsearch.spi.facet.FacetProcessor;
 
 import java.util.HashMap;
 import java.util.Map;
