+++ modules/portal/portal-search-elasticsearch/src/com/liferay/portal/search/elasticsearch/facet/DefaultFacetProcessor.java
@@ -16,7 +16,6 @@ package com.liferay.portal.search.elasticsearch.facet;
 
 import com.liferay.portal.kernel.search.facet.Facet;
 import com.liferay.portal.kernel.search.facet.config.FacetConfiguration;
-import com.liferay.portal.search.elasticsearch.spi.facet.FacetProcessor;
 
 import org.elasticsearch.action.search.SearchRequestBuilder;
 import org.elasticsearch.search.aggregations.bucket.terms.TermsBuilder;
