+++ modules/portal/portal-search-elasticsearch/src/com/liferay/portal/search/elasticsearch/query/BooleanQueryTranslatorImpl.java
@@ -19,7 +19,6 @@ import com.liferay.portal.kernel.search.BooleanClauseOccur;
 import com.liferay.portal.kernel.search.BooleanQuery;
 import com.liferay.portal.kernel.search.Query;
 import com.liferay.portal.kernel.search.QueryVisitor;
-import com.liferay.portal.search.elasticsearch.spi.query.BooleanQueryTranslator;
 
 import org.elasticsearch.index.query.BoolQueryBuilder;
 import org.elasticsearch.index.query.QueryBuilder;
