+++ modules/portal/portal-search-elasticsearch/src/com/liferay/portal/search/elasticsearch/query/TermQueryTranslatorImpl.java
@@ -19,7 +19,6 @@ import com.liferay.portal.kernel.search.QueryTerm;
 import com.liferay.portal.kernel.search.TermQuery;
 import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.kernel.util.StringUtil;
-import com.liferay.portal.search.elasticsearch.spi.query.TermQueryTranslator;
 
 import org.elasticsearch.index.query.QueryBuilder;
 import org.elasticsearch.index.query.QueryBuilders;
