+++ modules/portal/portal-search-elasticsearch/src/com/liferay/portal/search/elasticsearch/query/WildcardQueryTranslatorImpl.java
@@ -16,7 +16,6 @@ package com.liferay.portal.search.elasticsearch.query;
 
 import com.liferay.portal.kernel.search.QueryTerm;
 import com.liferay.portal.kernel.search.WildcardQuery;
-import com.liferay.portal.search.elasticsearch.spi.query.WildcardQueryTranslator;
 
 import org.elasticsearch.index.query.QueryBuilder;
 import org.elasticsearch.index.query.QueryBuilders;
